# index.js
const express = require('express')
const cors = require('cors')
const app = express()
const { MongoClient, ServerApiVersion, ObjectId } = require('mongodb');
require('dotenv').config();

//swagger ui express
const swaggerUi = require('swagger-ui-express');
const swaggerDocument = require('./swagger.json');

app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));


const port = process.env.PORT || 3000

//middleware
app.use(cors())
app.use(express.json())



// const uri = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASS}@cluster0.mrbmxyw.mongodb.net/?appName=Cluster0`;

const uri =`mongodb://localhost:27017`

// Create a MongoClient with a MongoClientOptions object to set the Stable API version
const client = new MongoClient(uri, {
  serverApi: {
    version: ServerApiVersion.v1,
    strict: true,
    deprecationErrors: true,
  }
});

app.get('/', (req, res) => {
  res.send('smart server running')
})

async function run() {
  try {

    // Connect to the "sample_mflix" database and access its "movies" collection
    const database = client.db("travelEase_db");
    const productCollection = database.collection("allVehicles");
    const usersCollection = database.collection("addVehicle")
    const bookNowCollection = database.collection("bookNow")
    //user collection
    const userCollection = database.collection('user')

    //user api
    app.post('/user', async(req, res)=>{
      const user = req.body
      user.createdAt = new Date().toISOString()
      const email = user.email;
      const userExists = await userCollection.findOne({email});

      if(userExists){
        return res.send({massage: 'User already exists'})
      }
      const result = await userCollection.insertOne(user)
      res.send(result)
    })

    // app.get('/user', async(req, res)=>{
    //   const email = req.query.email
    //     const query = {}
    //     if(email){
    //         query.currentUserEmail = email
    //     }
    //     const cursor = userCollection.find(query)
    //     const result = await cursor.toArray();
    //     res.send(result)

    // })

    app.get('/user', async(req, res)=>{
  const email = req.query.email;
  const result = await userCollection.find({ email }).toArray();
  res.send(result);
});

    app.get('/user/:id', async (req, res)=>{
        const id = req.params.id;
        const query = { _id: new ObjectId(id)}
        const result = await userCollection.findOne(query);
        res.send(result)
    })
app.patch('/user', async (req, res) => {
  const email = req.query.email;
  const { name, photo } = req.body;

  const result = await userCollection.updateOne(
    { email },
    { $set: { name, photo } }
  );

  res.send(result);
});





    //add vehicles
    app.post('/addVehicle', async(req, res)=>{
        const newUser = req.body;
        newUser.createdAt = new Date(); 
        const result = await usersCollection.insertOne(newUser);
        res.send(result);
    })

    //latest vehicles
    app.get('/latest-vehicles',async(req,res)=>{
        const cursor = productCollection.find().sort({createdAt: -1}).limit(8);
        const result = await cursor.toArray();
        res.send(result)
    })

    app.get('/latest-vehicles/:id', async (req, res)=>{
        const id = req.params.id;
        const query = { _id: new ObjectId(id)}
        const result = await productCollection.findOne(query);
        res.send(result)
    })

    // app.get('/latest-vehicles/:id',async(req,res)=>{
    //     // const cursor = productCollection.find().sort({createdAt: -1}).limit(6);
    //     // const result = await cursor.toArray();
    //     // res.send(result)
    //     const id = req.params.id;
    //     const query = { _id: new ObjectId(id)}
    //     const result = await productCollection.findOne(query);
    //     res.send(result)
    // })

    //show all vehicles
    app.post('/allVehicles', async (req, res)=>{
        const newProduct = req.body;
        const result = await productCollection.insertOne(newProduct);
        res.send(result)
    })

    //book now 
    app.post('/bookNow', async (req, res)=>{
      const bookNow = req.body;
      const result = await bookNowCollection.insertOne(bookNow)
      res.send(result)
    })
    //book now by user
    app.get('/bookNow', async (req, res)=>{
     const email = req.query.email
      const result = await bookNowCollection.find({bookerEmail : email}).toArray()
res.send(result)
    })

    app.get('/allVehicles', async (req, res)=>{

        const email = req.query.email
        const query = {}
        if(email){
            query.currentUserEmail = email
        }
        const cursor = productCollection.find(query)
        const result = await cursor.toArray();
        res.send(result)
    })

    app.get('/allVehicles/:id', async (req, res)=>{
        const id = req.params.id;
        const query = { _id: new ObjectId(id)}
        const result = await productCollection.findOne(query);
        res.send(result)
    })

    app.get('/myVehiclePage', async(req,res)=>{
      const email = req.query.email
      const result = await productCollection.find({currentUserEmail : email}).toArray()
res.send(result)
    })

    app.put('/allVehicles/:id', async(req, res)=>{
        const id = req.params.id;
        const UpdatedData = req.body
        const query = { _id: new ObjectId(id)}
        const update = {
            $set: UpdatedData
        }
        const result = await productCollection.updateOne(query, update);
        res.send(result)
    })

    app.patch('/allVehicles/:id', async(req, res)=>{
        const id = req.params.id;
        const updateProduct = req.body;
        const query = { _id : new ObjectId(id)}
        const update={
            $set: {
                name: updateProduct.name,
                price: updateProduct.price
            }
        }
            const result = await productCollection.updateOne(query, update);
            res.send(result)

    })

    app.delete('/allVehicles/:id', async(req, res)=>{
        const id = req.params.id;
        const query = { _id: new ObjectId(id)}
        const result = await productCollection.deleteOne(query);
        res.send(result)
    })
    


    // Connect the client to the server	(optional starting in v4.7)
    await client.connect();
    // Send a ping to confirm a successful connection
    // await client.db("admin").command({ ping: 1 });
    // console.log("Pinged your deployment. You successfully connected to MongoDB!");
  } finally {
    // Ensures that the client will close when you finish/error
    // await client.close();
  }
}
run().catch(console.dir);

app.listen(port, () => {
  console.log(`smart server running on port ${port}`)
})